requirements-dev.txt
-r requirements.txt

pytest==6.0.1
pytest-cov==2.10.1
pytest-django==3.9.0
pytest-mock==3.3.1
selenium==3.141.0
coverage==5.2.1
flake8==3.8.3
flake8-bugbear==20.1.4
black==19.10b0
isort==5.5.4
mypy==0.790
pre-commit == 2.7.1
tox == 3.21.2
